sugar/lift: model wide-int literals exactly (u64/u128/isize) — drain number-too-large
DIG the large-int-literal bucket: a wide Rust literal (u64::MAX, a u128 within
i128 range, a large isize) is a VALID literal and an EXACT mathematical-Int
constant — the FOL/SMT `Int` sort is unbounded, so there is no "too large" for
anything i128 can represent. Widen the integer carrier i64 -> i128 across the
IR / canonicalizer / proofir / SMT-compiler / lifter so these lift to their
EXACT value instead of refusing.

parse_int_lit parses via i128::from_str_radix; a literal beyond i128 range
(u128::MAX, the 128-bit step.rs / intrinsics.rs constants) still REFUSES
"number too large" — honest EXACT-OR-BAIL, never a silent truncation.

Precision-safe encoding (serde_json default has no arbitrary_precision; a bare
> u64 JSON number parses back to a lossy f64): an Int within i64/u64 stays a
JSON Number (byte-identical to the former i64 encoding — CID conserved for
every value the kit produced before); a value > u64 within i128 is carried as a
decimal STRING, disambiguated from a genuine String const by the `Int` sort.
The live SMT term renderer emits that string as its numeral; const_concrete_int
recovers it into the i128 distinctness set so no refutation is masked.

const_eval (the .map/.filter/.fold defolder) keeps width-checked i64 arithmetic
— rustc panics on i64::MAX*1000 in debug, so folding it in i128 would model a
computation the vendor's test never performs (a fake-dig). A wide value in
defold arithmetic BAILS; the `as i64`/`as isize` cast arm bails when the value
would truncate. Direct wide-literal assertions still lift via translate_lit.

Sweep (--rustc-cfg, hermetic): discharged 6188->6240 (+52), refused 687->662
(-25: 16 "number too large" reclassified 19->3, plus wide-int-gated wrapping/
step bodies now lift), unclassified 59->56 (-3: int-width-gated if/match bodies
drained), SILENT 0, assertion_multiset_cid ee0c6d92...0a50 CONSERVED. The 3
remaining "number too large" are genuinely > i128::MAX (honest bails). lib
130/0, assertion_lift 254/0 (+3 wide-int dig/twin/bail tests), workspace
--all-targets clean.

+++ implementations/rust/sugar-canonicalizer/src/value.rs
@@ -20,7 +20,12 @@ pub enum ValueKind {
 pub enum Value {
     Null,
     Bool(bool),
-    Integer(i64),
+    // Integer carrier is i128: a wide Rust literal (u64::MAX, u128 within
+    // i128 range, large isize) is an EXACT mathematical-Int constant. JCS
+    // emits plain decimal digits (`to_string`), so any i128 canonicalizes
+    // exactly and a small value's bytes are byte-identical to the former
+    // i64 form (CID conserved).
+    Integer(i128),
     String(String),
     Array(Vec<Arc<Value>>),
     Object(Vec<(String, Arc<Value>)>),
@@ -44,7 +49,7 @@ impl Value {
     pub fn boolean(b: bool) -> Arc<Value> {
         Arc::new(Value::Bool(b))
     }
-    pub fn integer(n: i64) -> Arc<Value> {
+    pub fn integer(n: i128) -> Arc<Value> {
         Arc::new(Value::Integer(n))
     }
     pub fn string<S: Into<String>>(s: S) -> Arc<Value> {
